requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
atila==0.26.11
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
colour==0.1.5
comm==0.1.3
contourpy==1.0.7
cuda-python==12.1.0
cv==1.0.0
cycler==0.11.0
Cython==0.29.34
dataclasses==0.6
datasets==2.10.1
debugpy==1.6.6
decorator==5.1.1
deepface==0.0.79
defusedxml==0.7.1
dill==0.3.6
dlib==19.24.1
dm-tree==0.1.8
dnn==0.7.4
dominate==2.7.0
efficientnet==1.0.0
entrypoints==0.4
executing==1.2.0
exif==1.6.0
fastapi==0.95.0
fastjsonschema==2.16.3
fdet==0.2.1
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gin-config==0.5.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.24.1
gradio_client==0.0.8
grpcio==1.53.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.4
image-classifiers==1.0.0
image-slicer==2.1.1
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.11.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vggface @ git+https://github.com/rcmalli/keras-vggface.git@719ad75625fe51436b9a9479114b3d136847e11a
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
lit==15.0.7
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.0.4
mpmath==1.2.1
mtcnn==0.1.1
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nbimporter==0.3.4
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==2022.4.8
nvidia-cublas-cu117==11.10.1.25
nvidia-cublas-cu12==12.1.0.26
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvcc==11.3.58
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.55
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.0.131
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
opencv-contrib-python==4.3.0.36
opencv-python==4.5.5.64
opencv-python-headless==3.4.18.65
opt-einsum==3.3.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
plum-py==0.8.6
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
python-version==0.0.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
replicate==0.5.2
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
retina-face==0.0.13
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rs4==0.3.23
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
segmentation-models==1.0.1
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.97
setproctitle==1.3.2
six==1.15.0
skitai==0.56.10
sniffio==1.3.0
soupsieve==2.4
sqlphile==0.9.7
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-model-optimization==0.7.4
tensorrt==8.6.0
tensorrt-dispatch==8.6.0
tensorrt-lean==8.6.0
termcolor==1.1.0
terminado==0.17.1
text-unidecode==1.3
tf-estimator-nightly==2.13.0.dev2023041108
tf-slim==1.1.0
tfserver==0.4.9
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1+cu118
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
ttictoc==0.5.6
typeguard==2.13.3
typing_extensions==4.5.0
uc-micro-py==1.0.1
Unidecode==1.3.6
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.25.11
uvicorn==0.21.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
webm==0.12.1
webrtcvad==2.0.10
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.12.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
